requirements.txt
paho-mqtt==1.6.1
pyinotify==0.9.6
python-dotenv==1.0.0
aiomqtt==1.2.1
